Add support for more script attributes
It's great that loadjs supports the `async` attribute but I need it to
support `crossorigin` in an app that I am building. In looking at the
[MDN:script](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/script)
docks I noticed that some people might want support for the integrity
tag as well.

I also found a nice way to test script tags for attributes by searching
for them based on the script src with a unique query paramter added.
Using this method might be easier than manually testing that
`async===false` does what `async===false` is supposed to do but instead
just trusting that if it was set the browser treats it as it should.
Not saying that needs to change, just an idea as it is very easy to
test for the attribute values using the method I used in the added
tests.