Update Readme to clarify usage of the library
I got confused for a while thinking that I have to include my app source code in the HTML for me to use load.js. I believe what the doc means is that I have to include load.js source in the HTML file and then load the app code using it.
Feel free to reject if I'm mistaken or if this change is unnecessary.